requirements.txt
sqlalchemy
psycopg2
flask
flask_restful
pixivpy-async
pixivpy
vk_api
requests
loguru
aiohttp_socks